Add files via upload
Alexis Perez-Bellido
First version of the task (not even piloted)
In this task the participants are presented with 10 posible stimuli positions. These positions are temporal associated in 5 pairs that then are presented as a continuous stream. I predict that participants will implicitly learn the associations and use them to improve the detection of trailing oddballs (radial gratings in this version) compared to leardin oddballs in the pairs.
In addition, the objective is also to test whether learning the associations has an effect on working memory: Participants have to use the mouse to report (after 3 s delay) the position of the last stimulis in the sequence (that sometimes will be a leading and others a trailing item).